+++ src/localExecution/IDafnyRunner.ts
@@ -1,4 +1,0 @@
-"use strict";
-export interface IDafnyRunner {
-  run(filename: string): void;
-}