+++ src/serverRequests/ICompile.ts
@@ -1,6 +1,5 @@
 "use strict";
-import { IDafnyRunner } from "../localExecution/_LocalExecutionModule";
+
 export interface ICompile {
-  compile(useCustomArgs: boolean): Promise<boolean>;
-  run(runner: IDafnyRunner): void;
+  compileAndRun(useCustomArgs: boolean, run: boolean): Promise<boolean>;
 }