+++ src/stringResources/commands.ts
@@ -20,7 +20,9 @@ export class Config {
   public static LanguageServerRuntimePath: string = "languageServerRuntimePath";
   public static LanguageServerLaunchArgs: string = "languageServerLaunchArgs";
   public static ColorCounterExamples: string = "colorCounterExamples";
-  public static CompilationArguments: string = "compilationArgs";
+  public static CompilerArgs: string = "compilerArgs";
+  public static CompilerRuntimePath: string = "compilerRuntimePath";
+  public static CompilerOutputDir: string = "compilerOutputDir";
 }
 
 export class EnvironmentConfig {